Remove SSH key passphrase test seam
