+++ app/src/main/java/com/pocketshell/app/hosts/SshKeysViewModel.kt
@@ -201,6 +201,4 @@ class SshKeysViewModel @Inject constructor(
         }
     }
 
-    internal fun hasPrivateKeyPassphrase(content: String): Boolean =
-        SshKeyStorage.hasPrivateKeyPassphrase(content)
 }
